Tidy up file structure and code
Previously, the Node and Tree classes were in a shared file because it
was easier to work with them in the same file when I was getting the
bones of the classes and the interactions between them set up.

Now, these classes have been split out into their own files. I have also
remove a bunch of redundant conditionals.