Refactor Tree class
Some improvements have been made to the `find` algorithm by adding a
couple of private helper methods - `_traverse_nodes` and
`_find_key_in_node`. I suspect that at some point in the future, the
tree traversal might live in a `TreeWalker` class, but this is a
sufficient improvement for the moment.